pyproject.toml
[tool.poetry]
name = "password-validator"
version = "0.1.0"
description = ""
authors = ["Pedro Varella <pvarellagomes@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = "^0.111.1"
pydantic-settings = "^2.3.4"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
python-multipart = "^0.0.9"
sqlalchemy = "^2.0.31"
authlib = "^1.3.1"
pytest-mock = "^3.14.0"
jwt = "^1.3.1"
python-jose = "^3.3.0"
pyjwt = "^2.8.0"
alembic = "^1.13.2"
uvicorn = "^0.30.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.4"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
run = 'fastapi dev src/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=src -vv'
post_test = 'coverage html'
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format' 

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select =['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
